Fix initial toggle only running when theme is set
When coming from an unset state or a light theme the switch would not
toggle according the state. So you had to toggle it once to light theme
and back again to dark.

This was an error because the check for the media preference for light
theme (unset) was included in the check for the localStorage key `theme`

Was fixed duplicating this toggle and run it if currentTheme is not set
and media preference matches light.